pdfs/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
asyncio==3.4.3
certifi==2024.6.2
distro==1.9.0
easyocr==1.7.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
fsspec==2024.6.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
imageio==2.34.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
openai==1.35.3
opencv-python-headless==4.10.0.84
packaging==24.1
pdf2image==1.17.0
pillow==10.3.0
pyclipper==1.3.0.post5
pydantic==2.7.4
pydantic_core==2.18.4
python-bidi==0.4.2
python-dotenv==1.0.1
PyYAML==6.0.1
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
sympy==1.12.1
tifffile==2024.6.18
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.12.2
